requirements.txt
adal>=0.4
requests==2.18.4
Flask==0.12.2
pyyaml>=3.0